Update Identifier logo subcomponent name to IdentifierLogo to differentiate between it and the Logo in Footer
