+++ src/components/Identifier/IdentifierLogo/IdentifierLogo.test.tsx
@@ -1,0 +1,15 @@
+import React from 'react'
+import { render } from '@testing-library/react'
+
+import { IdentifierLogo } from './IdentifierLogo'
+
+describe('IdentifierLogo component', () => {
+  it('renders without errors', () => {
+    //
+  })
+
+  it('renders multiple logos', () => {
+    // do I test this here or in Identifier.stories ?
+    //
+  })
+})
