--- src/components/Identifier/Logo/Logo.test.tsx
@@ -1,18 +1,0 @@
-import React from 'react'
-import { render } from '@testing-library/react'
-
-import { Logo } from './Logo'
-
-describe('Logo component', () => {
-  it('renders without errors', () => {
-    // const { getByRole } = render(<Logo />)
-    // expect(getByRole('anchor')).toBeInTheDocument
-  })
-  it('renders logo image', () => {
-    //
-  })
-  it('renders when multiple logos are passed in', () => {
-    // do I test this here or in Identifier.stories ?
-    //
-  })
-})
